Extract the missing documents logic on a separate controller
