+++ app/views/planning_applications/validation/documents/edit.html.erb
@@ -50,7 +50,7 @@
     </p>
     <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
 
-    <%= form_with model: @planning_application, url: validate_documents_planning_application_path(@planning_application) do |form| %>
+    <%= form_with model: @planning_application, url: planning_application_validation_documents_path(@planning_application) do |form| %>
       <%= form.hidden_field :documents_missing, value: @additional_document_validation_requests.any? %>
 
       <div class="govuk-button-group">
