+++ spec/components/task_list_items/additional_document_component_spec.rb
@@ -14,7 +14,7 @@ RSpec.describe TaskListItems::AdditionalDocumentComponent, type: :component do
   it "renders link" do
     expect(page).to have_link(
       "Check missing documents",
-      href: "/planning_applications/#{planning_application.id}/validation_documents"
+      href: "/planning_applications/#{planning_application.id}/validation/documents/edit"
     )
   end
 end
